apps/api/requirements.txt
# apps/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
transformers>=5.0.0
accelerate==1.0.1
torch
